+++ contracts/interfaces/IFolio.sol
@@ -130,7 +130,7 @@ interface IFolio {
         FeeRecipient[] immutableFeeRecipients;
         uint256 tvlFee; // D18{1/year} annual fee input; stored on Folio as D18{1/s}
         uint256 mintFee; // D18{1}
-        uint256 folioFeeForSelf; // D18{1} fraction of fee-recipient shares to burn
+        uint256 folioFeeForSelf; // D18{1} fraction of fee-recipient value directed to Folio holders
         string mandate;
     }